online-banking-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic[email]
passlib[bcrypt]
python-dotenv
PyJWT
